business logic for service layer

+++ src/main/java/com/surver/repositories/ParticipationRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ParticipationRepository extends JpaRepository<Participation,Integer> {
+public interface ParticipationRepository extends JpaRepository<Participation,Long> {
 }
